Address Kimi provider review: shared ACP bases + fixes
Review follow-ups on the Kimi provider (behavior-preserving except the
auth-message fix, which is the point):

Fixes:
- A logged-out Kimi (empty model option list) now fails session start and
  git text generation with the "run kimi login" message instead of an
  opaque client-side "expected one of <empty>" config-validation error.
- resolveKimiBinaryPath's PATH scan result is threaded through the driver
  so adapter/text-gen/status-probe spawns skip re-scanning.
- findKimiModelConfigOption matches by id only (like the generic write
  path); requiring the category tag could split-brain discovery vs writes.
- Narrowed the Kimi mapError method tag to the single RPC it uses.
- Added the missing kimi (and grok) custom-model input placeholders.

Refactors (byte-identical behavior; extractions the review called for):
- StandardAcpAdapter.ts: the provider-agnostic ACP adapter core, moved out
  of CursorAdapter.ts. Cursor's private extension protocol is now supplied
  through a generic registerExtensions hook instead of a Cursor-named
  boolean; Cursor and Kimi are thin wirings over the shared base.
- StandardAcpTextGeneration.ts: the shared ACP text-generation core, moved
  out of GrokTextGeneration.ts; Grok and Kimi delegate to it.
- providerStatusProbe.ts: the shared CLI status-probe + version-advisory
  scaffold, used by Kimi and Grok (Cursor's probe left as-is to avoid
  behavior risk). All per-provider message strings preserved exactly.
- AcpModelSelection.ts: shared applyAcpModelSelectionIfChanged, used by the
  Grok and Kimi model-selection helpers.

All 5 pre-existing Windows test failures (POSIX mock-binary spawns) are
unchanged; typecheck and vp check are clean.
